+++ service/cleaning/api/internal/logic/createorderlogic.go
@@ -142,7 +142,7 @@ func (l *CreateOrderLogic) CreateOrder(req *types.CreateOrderRequest) (resp *typ
 	}
 
 	// Property and region charge
-	pr_charge := float64(1.0 + (region_item.ChargeAmount+property_item.ChargeAmount)/100.0)
+	pr_charge := float64(1.0 + float64(region_item.ChargeAmount+property_item.ChargeAmount)/100.0)
 
 	// Service
 	// Basic Services
